move filter items to util
