+++ src/app/utils/itemsContainer.js
@@ -11,4 +11,31 @@ function chunk (arr, n) {
   return [arr.slice(0, n)].concat(chunk(arr.slice(n), n));
 }
 
-export { chunk }
+function filterItems (items, query, hasFilter) {
+  if (!items || !items.length) return [];
+  if (!query) return items;
+  if (!hasFilter) return items;
+
+  return items.filter((item) => {
+    const showItem = itemFilters.every((filter) => {
+      const filterType = filter.type;
+      const filterValue = query[filterType];
+      if (!filterValue) return true;
+      const selections =
+        typeof filterValue === 'string' ? [filterValue] : filterValue;
+      return selections.some((selection) => {
+        const isRequestable =
+          filterType === 'status' && selection === 'requestable';
+        if (isRequestable) return item.requestable;
+        const isOffsite =
+          filterType === 'location' && selection === 'offsite';
+        if (isOffsite) return item.isOffsite;
+        const itemProperty = filter.retrieveOption(item).label;
+        return isOptionSelected(selection, itemProperty, true);
+      });
+    });
+    return showItem;
+  });
+}
+
+export { chunk, filterItems }
